Completed draft for the Appointment class
This is the draft for the appointment class, it will be changed and adapted later to fit the project better, but for now this should serve as a good outline for the class. Btw, I made it so that the appointment cannot make an appointment within five days from the current date.